+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- changed: Randomize data-sync start times.
+
 ## 2.9.2 (2024-07-25)
 
 - fixed: Work with React Native bridgeless mode on iOS and Android.